added controller for ch01 CA manual
