+++ src/types/page.urls.ts
@@ -206,7 +206,7 @@ export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH_END = UPDATE_PAT
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH_END;
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_SEARCH;
 
-export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END = UPDATE_PATH + "director-correspondence-address-manual";
+export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END = UPDATE_PATH + "enter-director-correspondence-address";
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL = ACTIVE_SUBMISSION_BASE + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_END;
 export const UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL_PATH = OFFICER_FILING + UPDATE_DIRECTOR_CORRESPONDENCE_ADDRESS_MANUAL;
